requirements.txt
together==1.5.3
python-dotenv==1.1.0
pyyaml==6.0.2
pandas==2.2.3
numpy==1.26.4
pytest==8.3.5
scikit-learn==1.6.1
scipy==1.15.2
tqdm==4.67.1
pydantic==2.11.0
openai==1.70.0
httpx==0.28.1
pyarrow==19.0.1
typing-extensions==4.13.0
matplotlib==3.10.1
seaborn==0.13.2
requests==2.32.3
tabulate==0.9.0
joblib==1.4.2